REFACTOR: Changed onboarding and login/signup to detect where the user is coming from for redirection. If no search parameter is present redirect to a business list. At login onSuccess callback determine if business, if so retrieve slug and redirect to the admin portal. Change the customer layout to server component and added dynamic info based on slug
