+++ src/app/business-list/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <div>business list</div>;
+}
